+++ content/changelog.md
@@ -1,5 +1,10 @@
+# Jan 8, 2022 Part 2: Electric Boogaloo
+- Added fielding stats!
+    - There are some `Opp_1`, `Opp_2`, `Made_1`, etc, stats I have no idea what they're supposed to mean... If you know, tell me what they mean.
+- misc fixes and styling adjustments
+
 # Jan 8, 2022
-- Added "search" query to the batting-stats table. So now you can include searches in urls. ex: [/batting-stats?search=Mike+Cruz](/batting-stats?search=Mike+Cruz+overall)
+- Added "search" query to the batting-stats table. So now you can include searches in urls. ex: [/batting-stats?search=Mike+Cruz+overall](/batting-stats?search=Mike+Cruz+overall)
 - Added basic [PWA](https://web.dev/progressive-web-apps/) functionality (faster load times and some offline usability)
 
 # Dec 30, 2022